flight-service/requirements.txt
# Flight Service Requirements  
# 항공편 검색 서비스용 Python 패키지 목록

# Flask 웹 프레임워크
flask==2.3.3
flask-cors==4.0.0

# JWT 토큰 검증 (인증된 사용자)
PyJWT==2.8.0

# 데이터베이스 연결 (MySQL/RDS)
mysql-connector-python==8.1.0

# 환경변수 관리
python-dotenv==1.0.0


# Redis 캐싱 (항공편 검색 결과, ElastiCache 연결)
redis==4.6.0

# HTTP 요청 (외부 API 호출)
requests==2.31.0